requirements.txt
streamlit==1.22.0
pandas==1.5.3
openpyxl==3.1.2
anthropic==0.2.8
pyjwt==2.6.0
bcrypt==4.0.1